requirements.txt
click==6.6
enum-compat==0.0.2
enum34==1.1.6
Flask==0.11.1
Flask-SocketIO==2.8.2
greenlet==0.4.11
ibmiotf==0.2.6
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
paho-mqtt==1.2
pyserial==2.7
python-engineio==1.1.0
python-socketio==1.6.2
pytz==2016.10
requests==2.12.4
six==1.10.0
Unidecode==0.4.19
virtualenv==15.0.3
Werkzeug==0.11.11